feat: create App instance
