+++ lib/app/app.dart
@@ -1,0 +1,1 @@
+export 'view/app.dart';
