+++ lib/app/view/app.dart
@@ -1,0 +1,39 @@
+import 'package:api_client/api_client.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:github_repository/github_repository.dart';
+import 'package:github_search/home/view/home_page.dart';
+
+class App extends StatefulWidget {
+  const App({required ApiClient apiClient, super.key}) : _apiClient = apiClient;
+
+  final ApiClient _apiClient;
+
+  @override
+  AppState createState() => AppState();
+}
+
+class AppState extends State<App> {
+  @override
+  Widget build(BuildContext context) {
+    return RepositoryProvider(
+      create: (context) => GithubRepository(apiClient: widget._apiClient),
+      child: const AppView(),
+    );
+  }
+}
+
+class AppView extends StatelessWidget {
+  const AppView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const HomePage(),
+    );
+  }
+}
